Add pendingtx support to overview page
