+++ www/partials/asset_types.html
@@ -22,7 +22,11 @@
               <div class="asset-item col-xs-5 text-left number">
                 <span ng-if="currency.property_type != undefined && currency.property_type == 1" class="currency" title="{{currency.balance}}" ng-bind-html="currency.balance | numeraljs: '0,0'"></span>
                 <span ng-if="currency.property_type != undefined && currency.property_type == 2" class="currency" title="{{currency.balance}}" ng-bind-html="currency.balance | numeraljs: '0,0.00000000'"></span>
-                <span ng-if="currency.property_type == undefined" class="currency" title="{{currency.balance}}"><span ng-bind-html="currency.balance | numeraljs: '0,0.00000000'"></span></span>
+                <span ng-if="currency.property_type == undefined" class="currency" title="{{currency.balance}}"><span ng-bind-html="currency.balance | numeraljs: '0,0.00000000'"></span> </span>
+                  <span ng-show="currency.pendingneg<0" tooltip-append-to-body='true' tooltip="This balance reflects Omni pending transaction(s) with a total outgoing amount of {{currency.pendingneg*-1}} {{currency.symbol}}">
+                    <a href="/about/faq" target="_new"><span class="glyphicon glyphicon-export"></span></a></span>
+                  <span ng-show="currency.pendingpos>0" tooltip-append-to-body='true' tooltip="Oh Boy, Someone sent you something. When it confirms this balance will be updated">
+                    <br><small><a href="/about/faq" target="_new"><span class="glyphicon glyphicon-import"></span></a> (+<span ng-bind-html="currency.pendingpos | numeraljs: '0,0.00000000'"></span> )</small></span>
               </div>
               <div class="asset-item col-xs-3 text-right number" title="{{currency.value}}" ng-bind-html="currency.value | numeraljs: '$0,0.00':true">
               </div>
