+++ www/partials/currency_detail_modal.html
@@ -16,7 +16,12 @@
                     <td>{{item.address.address}}</td>
                    <!-- <td ng-show={{item.address.privkey}}>{{item.address.address}}<span class="glyphicon glyphicon-lock" tooltip="Your Wallet has the Private key for this address"></span></td> -->
 		    <td class="currency number" title="{{item.balance}}" ng-bind-html="item.balance | numeraljs: '0,0.00000000'"></td>
-		    <td class="currency number" title="{{item.value}}" ng-bind-html="item.value | numeraljs: '$0,0.00':true"></td>
+                   <td class="currency number" title="{{item.balance}}"><span ng-bind-html="item.balance | numeraljs: '0,0.00000000'"></span>
+                      <span ng-show="item.pendingneg<0" tooltip="This balance reflects Omni pending transaction(s) with a total outgoing amount of {{item.pendingneg*-1}}">
+                        <a href="/about/faq" target="_new"><span class="glyphicon glyphicon-export"></span></a></span>
+                      <span ng-show="item.pendingpos>0" tooltip="Oh Boy, Someone sent you something from Omni. When it confirms this balance will be updated">
+                        <br><small><a href="/about/faq" target="_new"><span class="glyphicon glyphicon-import"></span></a> (+<span ng-bind-html="item.pendingpos | numeraljs: '0,0.00000000'"></span> Unconfirmed)</small></span>
+                    </td>
 		  </tr>
 		</tbody>
 	</table>
